feat: ensure default cover photo works for proposal action events
